[autoscaler] cache stopped nodes, no screen on attach (#5741)

+++ python/ray/autoscaler/aws/node_provider.py
@@ -12,7 +12,8 @@ import botocore
 from botocore.config import Config
 
 from ray.autoscaler.node_provider import NodeProvider
-from ray.autoscaler.tags import TAG_RAY_CLUSTER_NAME, TAG_RAY_NODE_NAME
+from ray.autoscaler.tags import TAG_RAY_CLUSTER_NAME, TAG_RAY_NODE_NAME, \
+    TAG_RAY_LAUNCH_CONFIG, TAG_RAY_NODE_TYPE
 from ray.ray_constants import BOTO_MAX_RETRIES
 from ray.autoscaler.log_timer import LogTimer
 
@@ -41,6 +42,8 @@ class AWSNodeProvider(NodeProvider):
     def __init__(self, provider_config, cluster_name):
         NodeProvider.__init__(self, provider_config, cluster_name)
         config = Config(retries={"max_attempts": BOTO_MAX_RETRIES})
+        self.cache_stopped_nodes = provider_config.get("cache_stopped_nodes",
+                                                       True)
         self.ec2 = boto3.resource(
             "ec2", region_name=provider_config["region"], config=config)
 
@@ -61,7 +64,7 @@ class AWSNodeProvider(NodeProvider):
         self.cached_nodes = {}
 
     def _node_tag_update_loop(self):
-        """ Update the AWS tags for a cluster periodically.
+        """Update the AWS tags for a cluster periodically.
 
         The purpose of this loop is to avoid excessive EC2 calls when a large
         number of nodes are being launched simultaneously.
@@ -171,6 +174,55 @@ class AWSNodeProvider(NodeProvider):
             self.tag_cache_update_event.set()
 
     def create_node(self, node_config, tags, count):
+        # Try to reuse previously stopped nodes with compatible configs
+        if self.cache_stopped_nodes:
+            filters = [
+                {
+                    "Name": "instance-state-name",
+                    "Values": ["stopped", "stopping"],
+                },
+                {
+                    "Name": "tag:{}".format(TAG_RAY_CLUSTER_NAME),
+                    "Values": [self.cluster_name],
+                },
+                {
+                    "Name": "tag:{}".format(TAG_RAY_NODE_TYPE),
+                    "Values": [tags[TAG_RAY_NODE_TYPE]],
+                },
+                {
+                    "Name": "tag:{}".format(TAG_RAY_LAUNCH_CONFIG),
+                    "Values": [tags[TAG_RAY_LAUNCH_CONFIG]],
+                },
+            ]
+
+            reuse_nodes = list(
+                self.ec2.instances.filter(Filters=filters))[:count]
+            reuse_node_ids = [n.id for n in reuse_nodes]
+            if reuse_nodes:
+                logger.info("AWSNodeProvider: reusing instances {}. "
+                            "To disable reuse, set "
+                            "'cache_stopped_nodes: False' in the provider "
+                            "config.".format(reuse_node_ids))
+
+                for node in reuse_nodes:
+                    self.tag_cache[node.id] = from_aws_format(
+                        {x["Key"]: x["Value"]
+                         for x in node.tags})
+                    if node.state["Name"] == "stopping":
+                        logger.info("AWSNodeProvider: waiting for instance "
+                                    "{} to fully stop...".format(node.id))
+                        node.wait_until_stopped()
+
+                self.ec2.meta.client.start_instances(
+                    InstanceIds=reuse_node_ids)
+                for node_id in reuse_node_ids:
+                    self.set_node_tags(node_id, tags)
+                count -= len(reuse_node_ids)
+
+        if count:
+            self._create_node(node_config, tags, count)
+
+    def _create_node(self, node_config, tags, count):
         tags = to_aws_format(tags)
         conf = node_config.copy()
 
@@ -248,13 +300,43 @@ class AWSNodeProvider(NodeProvider):
 
     def terminate_node(self, node_id):
         node = self._get_cached_node(node_id)
-        node.terminate()
+        if self.cache_stopped_nodes:
+            if node.spot_instance_request_id:
+                logger.info(
+                    "AWSNodeProvider: terminating node {} (spot nodes cannot "
+                    "be stopped, only terminated)".format(node_id))
+                node.terminate()
+            else:
+                logger.info(
+                    "AWSNodeProvider: stopping node {}. To terminate nodes "
+                    "on stop, set 'cache_stopped_nodes: False' in the "
+                    "provider config.".format(node_id))
+                node.stop()
+        else:
+            node.terminate()
 
         self.tag_cache.pop(node_id, None)
         self.tag_cache_pending.pop(node_id, None)
 
     def terminate_nodes(self, node_ids):
-        self.ec2.meta.client.terminate_instances(InstanceIds=node_ids)
+        if not node_ids:
+            return
+
+        node0 = self._get_cached_node(node_ids[0])
+        if self.cache_stopped_nodes:
+            if node0.spot_instance_request_id:
+                logger.info(
+                    "AWSNodeProvider: terminating nodes {} (spot nodes cannot "
+                    "be stopped, only terminated)".format(node_ids))
+                self.ec2.meta.client.terminate_instances(InstanceIds=node_ids)
+            else:
+                logger.info(
+                    "AWSNodeProvider: stopping nodes {}. To terminate nodes "
+                    "on stop, set 'cache_stopped_nodes: False' in the "
+                    "provider config.".format(node_ids))
+                self.ec2.meta.client.stop_instances(InstanceIds=node_ids)
+        else:
+            self.ec2.meta.client.terminate_instances(InstanceIds=node_ids)
 
         for node_id in node_ids:
             self.tag_cache.pop(node_id, None)
